:art: 优化重构并统一公众号和小程序的spring boot starter部分配置类和属性
