+++ spring-boot-starters/wx-java-mp-spring-boot-starter/README.md
@@ -16,12 +16,12 @@
     wx.mp.token = @token
     wx.mp.aesKey = @aesKey
     # 存储配置redis(可选)
-    wx.mp.config-storage.type = redis                     # 配置类型: memory(默认), redis, jedis, redistemplate
+    wx.mp.config-storage.type = Jedis                     # 配置类型: Memory(默认), Jedis, RedisTemplate
     wx.mp.config-storage.key-prefix = wx                  # 相关redis前缀配置: wx(默认)
     wx.mp.config-storage.redis.host = 127.0.0.1
     wx.mp.config-storage.redis.port = 6379
     # http客户端配置
-    wx.mp.config-storage.http-client-type=httpclient      # http客户端类型: httpclient(默认), okhttp, joddhttp
+    wx.mp.config-storage.http-client-type=httpclient      # http客户端类型: HttpClient(默认), OkHttp, JoddHttp
     wx.mp.config-storage.http-proxy-host=
     wx.mp.config-storage.http-proxy-port=
     wx.mp.config-storage.http-proxy-username=
