+++ spring-boot-starters/wx-java-mp-spring-boot-starter/src/main/java/com/binarywang/spring/starter/wxjava/mp/enums/StorageType.java
@@ -1,0 +1,22 @@
+package com.binarywang.spring.starter.wxjava.mp.enums;
+
+/**
+ * storage类型.
+ *
+ * @author <a href="https://github.com/binarywang">Binary Wang</a>
+ * @date 2020-08-30
+ */
+public enum StorageType {
+  /**
+   * 内存.
+   */
+  Memory,
+  /**
+   * redis(JedisClient).
+   */
+  Jedis,
+  /**
+   * redis(RedisTemplate).
+   */
+  RedisTemplate
+}
