+++ spring-boot-starters/wx-java-mp-spring-boot-starter/src/main/java/com/binarywang/spring/starter/wxjava/mp/properties/WxMpProperties.java
@@ -1,12 +1,14 @@
 package com.binarywang.spring.starter.wxjava.mp.properties;
 
+import com.binarywang.spring.starter.wxjava.mp.enums.HttpClientType;
+import com.binarywang.spring.starter.wxjava.mp.enums.StorageType;
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 import java.io.Serializable;
 
 import static com.binarywang.spring.starter.wxjava.mp.properties.WxMpProperties.PREFIX;
-import static com.binarywang.spring.starter.wxjava.mp.properties.WxMpProperties.StorageType.memory;
+import static com.binarywang.spring.starter.wxjava.mp.enums.StorageType.Memory;
 
 
 /**
@@ -51,7 +53,7 @@ public class WxMpProperties {
     /**
      * 存储类型.
      */
-    private StorageType type = memory;
+    private StorageType type = Memory;
 
     /**
      * 指定key前缀.
@@ -88,75 +90,6 @@ public class WxMpProperties {
      */
     private String httpProxyPassword;
 
-  }
-
-  public enum StorageType {
-    /**
-     * 内存.
-     */
-    memory,
-    /**
-     * jedis.
-     */
-    redis,
-    /**
-     * redis(JedisClient).
-     */
-    jedis,
-    /**
-     * redis(RedisTemplate).
-     */
-    redistemplate
-  }
-
-  public enum HttpClientType {
-    /**
-     * HttpClient.
-     */
-    httpclient,
-    /**
-     * OkHttp.
-     */
-    okhttp,
-    /**
-     * JoddHttp.
-     */
-    joddhttp
-  }
-
-  @Data
-  public static class RedisProperties implements Serializable {
-    private static final long serialVersionUID = -5924815351660074401L;
-
-    /**
-     * 主机地址.
-     */
-    private String host = "127.0.0.1";
-
-    /**
-     * 端口号.
-     */
-    private int port = 6379;
-
-    /**
-     * 密码.
-     */
-    private String password;
-
-    /**
-     * 超时.
-     */
-    private int timeout = 2000;
-
-    /**
-     * 数据库.
-     */
-    private int database = 0;
-
-    private Integer maxActive;
-    private Integer maxIdle;
-    private Integer maxWaitMillis;
-    private Integer minIdle;
   }
 
 }
